requirements.txt
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
firebase-admin==5.4.0
Flask==2.2.2
google-api-core==2.10.1
google-api-python-client==2.62.0
google-auth==2.11.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-recaptcha-enterprise==1.8.1
google-cloud-secret-manager==2.12.4
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.49.1
grpcio-status==1.49.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
proto-plus==1.22.1
protobuf==4.21.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.2
